Correctly extract source maps from CoffeeScript.compile
When CoffeeScript.compile is normally called, the result is a string containing
the compiled js.  When sourceMap is set to true though, the result is a hash
with keys ['js', 'sourceMap', 'v3SourceMap'].

Before this change, we'd write that entire hash into the .map file, and I'm
somewhat sure that isn't a valid sourceMap.