Set source map file properties, and allow sourceRoot to be overridden.
The coffeescript compiler accepts various file and path options to help
the browser find the original source files that the map refers to.

sourceFiles - name of the original sources.

sourceRoot - Root url the browser prepends to all sourceFiles. It's useful to
override this if you want to change where/how these files are served, like adding
authentication.

generatedFile - I'm actually not sure why this is necessary, since the generated
file itself should refer to the map. I added in an effort to closely match the
output of the coffee binary.